requirements.txt
ollama
PyYAML
mysql-connector-python